git-svn-id: http://svn.openmrs.org/openmrs/trunk@1776 5bac5841-c719-aa4e-b3fe-cce5062f897a

+++ src/web/org/openmrs/web/controller/report/export/DataExportListController.java
@@ -1,7 +1,11 @@
 package org.openmrs.web.controller.report.export;
 
-import java.util.HashSet;
-import java.util.Set;
+import java.io.File;
+import java.util.Date;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Vector;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
@@ -16,10 +20,12 @@ import org.openmrs.api.context.Context;
 import org.openmrs.reporting.AbstractReportObject;
 import org.openmrs.reporting.ReportService;
 import org.openmrs.reporting.export.DataExportReportObject;
+import org.openmrs.reporting.export.DataExportUtil;
 import org.openmrs.web.WebConstants;
 import org.springframework.beans.propertyeditors.CustomNumberEditor;
 import org.springframework.context.support.MessageSourceAccessor;
 import org.springframework.validation.BindException;
+import org.springframework.validation.Errors;
 import org.springframework.web.bind.ServletRequestDataBinder;
 import org.springframework.web.servlet.ModelAndView;
 import org.springframework.web.servlet.mvc.SimpleFormController;
@@ -53,11 +59,13 @@ public class DataExportListController extends SimpleFormController {
 	protected ModelAndView onSubmit(HttpServletRequest request, HttpServletResponse response, Object obj, BindException errors) throws Exception {
 		
 		HttpSession httpSession = request.getSession();
-		Context context = (Context) httpSession.getAttribute(WebConstants.OPENMRS_CONTEXT_HTTPSESSION_ATTR);
+		
 		String view = getFormView();
-		if (context != null && context.isAuthenticated()) {
+		if (Context.isAuthenticated()) {
 			String[] reportList = request.getParameterValues("dataExportId");
-			AdministrationService as = context.getAdministrationService();
+			String action = request.getParameter("action");
+			
+			AdministrationService as = Context.getAdministrationService();
 			
 			String success = "";
 			String error = "";
@@ -67,23 +75,56 @@ public class DataExportListController extends SimpleFormController {
 			String notDeleted = msa.getMessage("general.cannot.delete");
 			String textDataExport = msa.getMessage("DataExport.dataExport");
 			String noneDeleted = msa.getMessage("DataExport.nonedeleted");
-			if ( reportList != null ) {
-				for (String p : reportList) {
-					//TODO convenience method deleteDataExport(Integer) ??
-					try {
-						as.deleteReportObject(Integer.valueOf(p));
-						if (!success.equals("")) success += "<br>";
-						success += textDataExport + " " + p + " " + deleted;
-					}
-					catch (APIException e) {
-						log.warn(e);
-						if (!error.equals("")) error += "<br>";
-						error += textDataExport + " " + p + " " + notDeleted;
+			
+			String generated = msa.getMessage("DataExport.generated");
+			String notGenerated = msa.getMessage("DataExport.notGenerated");
+			String noneGenerated = msa.getMessage("DataExport.noneGenerated");
+			
+			if (msa.getMessage("DataExport.generate").equals(action)) {
+				if (reportList == null)
+					success = noneGenerated;
+				else {
+					ReportService rs = Context.getReportService();
+					for (String id : reportList) {
+						DataExportReportObject report = null;
+						try {
+							report = (DataExportReportObject)rs.getReportObject(Integer.valueOf(id));
+							DataExportUtil.generateExport(report, null);
+							if (!success.equals("")) success += "<br>";
+							success += textDataExport + " '" + report.getName() + "' " + generated;
+						}
+						catch (Exception e) {
+							log.warn("Error generating report object", e);
+							if (!error.equals("")) error += "<br>";
+							if (report == null)
+								error += textDataExport + " #" + id + " " + notGenerated;
+							else
+								error += textDataExport + " '" + report.getName() + "' " + notGenerated;
+						}
 					}
 				}
-			} else {
-				success += noneDeleted;
 			}
+			else if (msa.getMessage("DataExport.delete").equals(action)) {
+				
+				if ( reportList != null ) {
+					for (String p : reportList) {
+						//TODO convenience method deleteDataExport(Integer) ??
+						try {
+							as.deleteReportObject(Integer.valueOf(p));
+							if (!success.equals("")) success += "<br>";
+							success += textDataExport + " " + p + " " + deleted;
+						}
+						catch (APIException e) {
+							log.warn("Error deleting report object", e);
+							if (!error.equals("")) error += "<br>";
+							error += textDataExport + " " + p + " " + notDeleted;
+						}
+					}
+				} else {
+					success += noneDeleted;
+				}
+			}
+			
 			view = getSuccessView();
 			if (!success.equals(""))
 				httpSession.setAttribute(WebConstants.OPENMRS_MSG_ATTR, success);
@@ -103,20 +144,57 @@ public class DataExportListController extends SimpleFormController {
 	 */
     protected Object formBackingObject(HttpServletRequest request) throws ServletException {
 
-    	HttpSession httpSession = request.getSession();
-		Context context = (Context) httpSession.getAttribute(WebConstants.OPENMRS_CONTEXT_HTTPSESSION_ATTR);
-		
 		//default empty Object
-		Set<AbstractReportObject> reportList = new HashSet<AbstractReportObject>();
+		List<AbstractReportObject> reportList = new Vector<AbstractReportObject>();
 		
 		//only fill the Object is the user has authenticated properly
-		if (context != null && context.isAuthenticated()) {
-			ReportService rs = context.getReportService();
+		if (Context.isAuthenticated()) {
+			ReportService rs = Context.getReportService();
 			//ReportService rs = new TestReportService();
 	    	reportList = rs.getReportObjectsByType(DataExportReportObject.TYPE_NAME);
 		}
     	
         return reportList;
     }
+
+	/**
+	 * @see org.springframework.web.servlet.mvc.SimpleFormController#referenceData(javax.servlet.http.HttpServletRequest, java.lang.Object, org.springframework.validation.Errors)
+	 */
+    @Override
+	@SuppressWarnings("unchecked")
+	protected Map referenceData(HttpServletRequest request, Object command, Errors errors) throws Exception {
+		
+		Map<String, Object> map = new HashMap<String, Object>();
+		
+		List<AbstractReportObject> reportList = (List<AbstractReportObject>) command;
+		Map<AbstractReportObject, Date> generatedDates = new HashMap<AbstractReportObject, Date>();
+		Map<AbstractReportObject, String> generatedSizes = new HashMap<AbstractReportObject, String>();
+		
+		// add the last modified date of the generated file as reference data
+		for (AbstractReportObject report : reportList) {
+			File file = DataExportUtil.getGeneratedFile((DataExportReportObject) report);
+			
+			if (file.exists()) {
+				generatedDates.put(report, new Date(file.lastModified()));
+				
+				Long size = file.length(); //returned in bytes
+				if (size > 1024*1024)
+					generatedSizes.put(report, size/(1024*1024) + "MB");
+				else if (size > 1024)
+					generatedSizes.put(report, size/1024 + "kB");
+				else 
+					generatedSizes.put(report, size + "B");
+			}
+		}
+		
+		map.put("generatedDates", generatedDates);
+		map.put("generatedSizes", generatedSizes);
+		
+		return map;
+		
+	}
+
+	
+    
     
 }
